Update k8s manifests - new image repos and no custom db image
The move to use kustomize is to prevent the need to create an
image specifically for the database. Since the file can simply be
mounted in, we can use the upstream image directly and mount it in.
By using kustomize, we can populate a ConfigMap with the same file
being mounted in the docker-compose.yml, reducing duplication.

Signed-off-by: Michael Irwin <mikesir87@gmail.com>